pyproject.toml
[tool.poetry]
name = "jarvis"
version = "0.1.0"
description = ""
license = "Apache License 2.0"
authors = ["Leal <douglas.sleal@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
databricks = "^0.2"
rich = "^13.7.1"
typer = "^0.12.3"
databricks-sdk = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.12.2"
databricks = "^0.2"
databricks-sdk = "^0.27.0"
pytz = "^2024.1"
azure-identity = "^1.17.1"
azure-mgmt-eventhub = "^11.0.0"
azure-mgmt-storage = "^21.2.1"
azure-mgmt-resource = "^23.1.1"
azure-mgmt-databricks = "^2.0.0"
faker = "^28.1.0"
datacontract-cli = "^0.10.11"
deltalake = "^0.19.1"
pyspark = "^3.5.2"
soda-core-spark-df = "^3.3.18"



[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.21"
mkdocstrings-python = "^1.10.0"
mkdocs-macros-plugin = "^1.0.5"
jinja2 = "^3.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=jarvis -vv"
post_test = "coverage html"
